Don't load Bundler from RubyGems tests
